typing: add a few hints to `mercurial.phases` to appease PyCharm
I'm a little surprised that pytype doesn't complain- `phasenumber` is
`dict[bytes, int]`, so that's what `phasenumber2` is, due to initializing with a
copy of `phasenumber`.  But then the next line doesn't adhere to that, and
starts adding keys as bytes.  Be explicit and allow both.

The second issue is PyCharm thinks `_readroots()` returns
`tuple[dict[int, set] | dict[int, set[int]], bool] | None` (note the first set
doesn't have a parameterized type like the second and the bool is somehow
missing).  That can also be helped by being explicit.  This doesn't completely
fix the issue (it still thinks it can return None somehow, but neither I nor
pytype see how, so it's got to be a bug), but at least the dict types now match
and are elided in the warning so that it is more readable.

--HG--
extra : rebase_source : c8920a571d05d8bf8eac00a39ffa12d38215968f
extra : topic : pytype-2024.10.11