setup-dev-tools: add make target to format the current stack using fix
Same as the "format" target, but for the whole stack.

In the current state of `hg fix` this leave the working copy behind and we
should probably fix that. This is out of scope for this series however.

--HG--
extra : rebase_source : 712ee0658fc090ba135dbd839aece63c2f4cef39
extra : topic : setup-dev-tools