makefile: more fix to the pip invocation
The Makefile is now usable by Mercurial developers and for downstream packaging.
`make help` explains how.

I had to introduce more options (PIP_OPTIONS_INSTALL, PIP_PREFIX) to be
able to change the behavior of the install targets (usable for local installation
and downstream packaging) and to deal with an issue on Linux with the --prefix
pip option (see
https://discuss.python.org/t/linux-distro-patches-to-sysconfig-are-changing-pip-install-prefix-outside-virtual-environments/18240).