dirstate-race: fix a missing synchronisation in the python code
This was giving the wrong output in the `dirstate-v2-rewrite` case. As the
dirstate is rewritten, the data file is missing and we reload the full dirstate.
The make the dirstate reloaded after the commit and it no longer see the file as
dirty.

Note that if we were deleting the data file less aggressively, we would see the other
output. (for example if we were keeping old data file around for some time
before deletion).

One last thing… fixing this reveal a quite serious bug where a `hg status`
end up overwriting a `hg update` done concurrently. Hooray for proper testing.