dirstate-race: simplify some output match to highligh an error
`#if dirstate-v2` is always false, because `dirstate-v2` is not one of the test
cases. So we fix the conditionnal and adjust the expected matching.

This highlight that the rhg/no-rhg output differ because an issue in the code.
We will fix that in the next changeset.

--HG--
branch : stable
extra : rebase_source : 4105c3ba55f959a424b07e039c129de8f1c287c2
extra : topic : dirstate-v2-sync-point.