dirstatemap: stop setting identity after reading the data
The code was calling `_set_identity` in too many places, resulting in fetching
cache information after reading the on disk data. This could result in old data
labelled with the identity of new, never loaded data.

Fixing it solve the issue spotted in the previous changesets.

Further cleanup are warranted but it seems more reasonable to do them on stable.