version: update the generation to something closer to what we had previously
When we moved to use setuptools-scm in 30510238284e, we lost the version scheme
that was engineers to make it safe to produce and consume nightly build from
multiple branches (stable, default, and privates branches).

This small change give use something close to what we have, but still don't
support multiple branch.