stream-clone: adjust default value for worker and memory
Increasing the memory usage does not seems to yield much speed up, so we can
restrict it more aggressively. However, unlimited memory usage can overwhelm a
system and result in massive slowdown if the process start swapping. So it is
still possible to configure it, but it is not longer used in any default
profile.

Having too many workers can slow things a bit, but not too much. So we can try
to use higher default to yield benefit on large system.

--HG--
extra : rebase_source : 9637b989893e7825314a34212d963f383bef51c9
extra : topic : streambundle-speed