test-fix-path: avoid a test hang on Windows
Windows can't typically invoke `*.py` directly as a command, and will instead
show a prompt asking what program should be used to open the file.  We can't
directly invoke the interpreter as is usually done in this case, because the
whole point is to run something not in `PATH`.  The easiest thing to do is
invoke a *.bat file that runs the interpreter.  We can get away with this
because the current implementation is to effectively run `cmd.exe /c %command%`,
which doesn't need the file extension specified.

--HG--
extra : amend_source : 694d25d716d66ccf36dd5b87c8d879e483bdd2c1
extra : topic : test-fix-path-windows