changegroup: directly use the InboundRevision attributes in remotefilelog
This make the code simpler. We are doing this change in a dedicated simpler to
reduce the noise in the initial changeset.

We also add a XXX pointing out that the "flag" from the inbound input end use being ignored and that is suspicious.

--HG--
extra : rebase_source : b9a82c8fb6984c8cb3413b4b3338cae5276bdef1
extra : topic : cg-flag-pre