setup: drop the inaccessible `py2exehacked` code around distutils
The comment heavily implies, and experimenting confirms, that there's no
`sys.real_prefix` in a venv on Python3.  I have no idea how the problems
described would manifest, but nobody has complained for the several years of
py3 releases build from a venv.  The experimenting was done with 6.9.2 and the
TortoiseHg build script.