annotate: use exit code 255 for missing files
This makes hg annotate always use exit code 255 for missing files. Before, it
used 10 (input error) if the file was never in the repo, and 255 (abort) if it
is in the repo for some other revision.

The motivation for this is to match the behavior of hg.py and rhg. This
discrepancy was revealed by using "hg annotate" instead of "hg ann" in
test-annotate.t, since rhg doesn't recognize "ann" and falls back.

--HG--
extra : rebase_source : ca7fe70c7b1ccc2195c4215677cee74f322079d7
extra : topic : annotate-wdir