stream-bundle: properly process requirements
When the manifest bundle constain some known requirement that does not affect
the stream clone, we used to crash. However since we know them and know they
don't affect the stream clone, we can ignore them.

Mozilla generated such buggy manifest bundle for a time which allowed us to
catch this error. The issue was not caught until 961900fbd67c (released in
6.9.2) as the requirements information were ignored for stream-v2 until then.

We fix the issue, refactor the code for robustness and adds more tests to better
catch this kind of issue in the future.

--HG--
branch : stable
extra : fix_source : 1580003141e218326b2b8cca2c3dc347195c4500
extra : topic : clonebundle-regression