nightly-build: attempt to no longer choke on existing file
When we run the job multiple time, the upload step fails because the file always
exists. So we pass the `--skip-existing` flag that is supposed to help with
that… when supported.

This might have no effect outside of upload to PyPI as the doc says:

    Continue uploading files if one already exists. (Only valid when uploading
    to PyPI. Other implementations may not support this.)

--HG--
branch : stable
extra : rebase_source : 16db28bb19cf0cb16157dbe0f637d84dc8b95160
extra : topic : fix-nighlty-upload