stream: audit path on encoded path
In 5b8f6e198a6e, we wrongly called `vfs._auditpath` on the pre-encoding path, so
we would get abort on problem that the encoding is here to handle.

So we fix the problem (easily) and update the test to cover this (with
significantly more work).

Thanks to Matt Harbison for spotting this during the freeze.

--HG--
branch : stable
extra : topic : fix-stream-windows