rust-bindings: don't change signals if not in the main thread (issue6961)
There is no way of checking if "in the main thread" in Rust, since there is
no such portable API on any OS and is part of the arbitrary runtime of any
language that wants to tag their main thread somehow. Thus, we ask Python
if we're in the main thread, and run the function with no sigint wrapping
otherwise.

We should also technically check for being in a subinterpreter, but I didn't
find a quick and obvious way of doing so and I'm pretty sure nobody cares.
If anyone cares in the future, we can address this.

--HG--
branch : stable
extra : rebase_source : cbbd6ddaf07ba3d0a2120aba6495a092745130a0
extra : topic : rust-signal-threading