rust-matchers: generalize consumers of the Matcher api somewhat
In the interest of making it easier to use the matchers API,
be more generous when producing matchers,
and be more permissive when consuming them.

- Add the `Send` trait everywhere where we're returning a Boxed matcher,
so the caller can be free to wrap it into an `Arc` to share between threads.

- For all matcher consumers, accept an &impl instead of a dyn box so
the caller is free to pass anything they want.

--HG--
extra : rebase_source : ca11cc54adb05f0abbea9fb369ce6c567a3ae7da
extra : topic : generalize-matchers