ci: print some extra info about $PYTHON when building wheels on macOS
I forgot how this was setup, and we don't seem to be using the venv that I
thought we were.

No, I don't know why these seem to need to be wrapped in `sh -c ''`.  But without
it, the pipeline was failing immediately with a syntax error, pointing to the
row and column of the first '-' and saying it couldn't find the '-' when parsing
the block.

--HG--
branch : stable
extra : rebase_source : ea28d155ace987c24bd723d5ee416f28afa71cac
extra : topic : ci-hacking