rust-config: preserve insertion order
This changes ConfigItem to be an indexmap::IndexMap instead of a HashMap, so
that it preserves insertion order. It also changes Config::iter_section to
return keys in that order (lowest to highest precedence). For all existing uses
of iter_section, the order doesn't appear to matter.

This is motivated by a follow-up change that will implement support for aliases
in rhg, where the order of alias definitions matters.

--HG--
extra : rebase_source : 7a7e677f5fd920afe1049e8cc6c9cea202315435
extra : topic : rhg-alias-support