ci: cache pytype inference and reuse it
We have to adjust the timestamp to set `ninja` on the right track.  The cache is
only generated for commit on the default branch to make sure it use a commit
that is known to other run (and quite recently generated).

--HG--
extra : rebase_source : f1c7c39d2f680bd7ca04c435c011f3bd3556fb1f
extra : topic : pytype-cache