branchmap: make sure we update the "hasnode" function on update
Without this, the  "hasnode" function could lag behind the actual content of the
branchmap. And while this was fine in most case. It was possible to craft case
were the branchhead had to be validated after their content were updated and it
would cause crash.

We should also close the gap on such scenarion, however not keeping an outdated
attribute around seems like a good idea.

--HG--
extra : topic : bm._hasnode