ci: use "v3.0" image for the for pycompat jobs
Same rational for the other switch. We now build multiple images based on the
official "python:3.<min-version>-alpine3.21" images.

We don't do it for Python 3.8 as it is EOL and the PSF no longer provides images
for it. Since we are looking into dropping support for 3.8, that seems fine to
not make further effort for it.

--HG--
extra : rebase_source : df39c0a9de24608f818e798f68171951acaa2345
extra : topic : ci-test-pys