branch: pass correct set of files to memctx() in case of merge (issue6972)
Previously, ctx.files() was passed (ctx refers to the changeset to be changed), which doesn’t return all relevant files in case of a merge. Instead, we use the same logic as in cmdutil.amend() and in the analogous code in the topic extension that changes the topic name.

--HG--
branch : stable
extra : rebase_source : 35b2e8648014152c47071585fd5587b96fa2a900
extra : topic : issue6972