procutil: actually silence backgrounded process stdio on Windows when asked
The code for using `None` as stdio descriptors to signal that the child process
should send output to `/dev/null` was added almost 5 years ago in 8759e22f1649.
Maybe this is a bad API, and `subprocess.DEVNULL` should be used, since `None`
seems to mean the child inherits the parent's descriptor.  Regardless, the child
was getting the parent descriptor because the Windows side of things wasn't
handled, and this failure was occurring in a Windows Server 2025 container:

    --- C:/hgdev/src/tests/test-clonebundles-autogen.t
    +++ C:/hgdev/src/tests/test-clonebundles-autogen.t.err
    @@ -444,7 +444,9 @@
       >     --config devel.clonebundles.override-operation-id=background_testing

       $ $RUNTESTDIR/testlib/wait-on-file 30 $v1_file
    +  11 changesets found
       $ $RUNTESTDIR/testlib/wait-on-file 30 $v2_file
    +  11 changesets found

     all file created, but some cleanup might still be in progress, we wait 30 second for them


    ERROR: test-clonebundles-autogen.t output changed

I've not seen it happen on bare metal Windows 11, nor on Windows 10 (but didn't
look through CI failures too hard for it).  Regardless, if the `--background`
arg was dropped from the test command on Windows 11, it emitted the same output,
so it's not a mystery where this comes from.

The code here is copy/pasted from the POSIX side (and thus the first conditional
is altered).  The `finally` block is also copied from the POSIX side to appease
pytype.

--HG--
branch : stable
extra : rebase_source : 79cdc44f7bb1ed4e50a891eb35255a5280f8509b
extra : topic : close-backgrounded-proc-stdio