debugindexdot: add flags for more control over graph
This adds some new flags to hg debugindexdot:

* -i, --include STATEMENT: arbitrary graphviz statement to include
* -r, --rev REV: only show these revisions
* --common: also show common ancestors of --rev
* -T, --template TEMPLATE: label template (only works on changelog)

If you give revisions with -r, it produces a graph like this:

* solid edge    -- parent -> child
* dashed edge   -- ancestor -> descendant (not implied by other edges)

If you add --common, then:

* unfilled node -- the revs given
* filled node   -- common ancestor of other revs

This is useful when debugging to quickly get a sense of how a set of revisions
relate to one another in the overall graph. And unlike log --graph, it can be
used on any revlog, not just the changelog.

--HG--
extra : rebase_source : 70157a87e4403abc13bd6ef7c02f0241bde79e12
extra : topic : debugindexdot-paths