run-tests: create the venv where hg is installed without `pip`
Otherwise, the test venv ends up with the default `pip` 22.0.4 and `setuptools`
58.1.0 when run with Python 3.9, which is too old to build Mercurial, and the
test run fails during the editable installation phase.  This happens even when
the base Python install has been updated to have the latest versions of these
packages.  This started out using `--upgrade-deps`, because I assume that
missing here was an oversight (`make local` upgrades the dependencies in the
venv it creates).  But this avoids some extra network fetching, and output
changes in `test-run-tests.t`.  We don't see this in CI because it's not an
issue when it is fed a wheel.

This also fixes the issue that we have seen in CI, where there is no
`--config-settings` option on `pip` on the default branch.

--HG--
branch : stable
extra : amend_source : 4d69787a50334e65edf2079088bfde4f665130bd
extra : topic : runtests-fixes