delta-search: start using delta fold to estimate delta size when available
In the simplest case, we are requested to compute a delta against the base of
the delta we just computed. In such case, we can use "delta folding" to estimate
and upper bound for the resulting delta.

This case notably happens when we try to refine a snapshot base down.

There is much more to do with delta folding, but this is a good start.

--HG--
extra : rebase_source : 1831eeb2d99f9df9468fc36377830bc81bcd3400
extra : topic : delta-search.fold-down