upgrade: allow to downgrade and upgrade from dotencode
The upgrade process rely on the store, but nothing actually prevent us to drop
the dotencode usage.

It is currently quite a bad idea, but when proper alternative gets in, that will
come handy.

In the short term, this will help to upgrade to and from the "plain encoding".

--HG--
extra : rebase_source : 676133bf134240cc5c7e045b9375b6b81747539b
extra : topic : plain_encoding