rust-vfs: add support for changing encodings
This is needed for the next changeset that will introduce plain encoding.

Our VFS story is not great yet: we have the `Vfs` trait that is missing some
functionality, the rest of which is hardcoded in `VfsImpl`, and we also have
the `FnCacheVfs` which is only useful from Python...

This cleanup will need to happen at some point, and will become necessary
when we get the `FileIndex`.

For now, I am making the code even worse, but not by so much that it feels
unreasonable considering all that needs to be cleaned up on our current
trajectory anyway.

--HG--
extra : rebase_source : 0c44e05d0b6bde90f3019796b6c1d3a64ab1e3a0
extra : topic : plain_encoding