fix: stop adding unexpected extra requirements to stream clones
The expected behavior is that stream clone inherits certain aspects
of the repo being cloned, in particular things like compression.

What we observe is that if a requirement is enabled by default,
it's also enabled in streaming clones, despite it not being
enabled in the source repo.
It's arguable whether or not that's good, but they way the code
is written makes it obvious this is a mistake.

--HG--
branch : stable
extra : rebase_source : 7053fd1871c2d90526e0520c3ecbe4d148b46665
extra : topic : fix-stream-clone-requirements-bug