apache-server: simplify the setup by installing Mercurial from PyPI
Instead of installing hg from source directory mounted as a Docker volume, we
now simply use PyPI to install whatever is available there (e.g. the latest
release, 7.0.2).

As long as we solve the "python packaging" issues somewhere in a better place,
and can successfully build and ship binary distributions of Mercurial, we
should be able to reuse the results of that work here and simplify everything.

Since at the moment we have binary wheels on PyPI and we don't need to build
anything inside the container, we can remove the development packages.

We can also stop installing recommends. With this patch the image went from
around 700MB to around 350MB.

This patch definitely changes the behavior of the image, however I think it
keeps the main idea intact: to just run hgweb in a production-like environment
and see how different configuration could affect it.

Keep in mind that even before this patch it was awkward to see any changes from
local working directory propagate into the hgweb inside the container, because
changing source code on host machine (by hand, for example) required restarting
the container to rebuild/reinstall Mercurial inside.

Okay, right now you can't see any source code modifications from the host
inside the container at all, unless the code gets published on PyPI, but I
think this is a fair compromise. This will allow this setup to more easily grow
other, different features. And, of course, if you want to simply see any of
your local changes in hgweb, you can run it directly, without setting up any
web servers and without using any containers.

I have chosen to pin Mercurial version explicitly and not just install the
latest release unconditionally hoping to keep the setup working in future for
as long as possible. The previous generation of apache-server (Debian Wheezy
and Python2) was fine for some time, but then we dropped support for Python2 in
newer releases and switched to modern packaging tools, and so the setup was
completely broken. Now we have pinned Debian release to Bookworm, and Python
version to 3.11, and can guarantee only specific versions of Mercurial to work
on that software stack without requiring any changes. But when the time comes
to update this setup again, at least we'll have something that works and can
serve as an example of what to expect after the update.

--HG--
extra : rebase_source : 522e29ff4fd9a33ba55f32ee1809d55d2fcf0ff8
extra : topic : docker-hg-apache