apache-server: download and install Anubis 1.19.1
Docker documentation suggests that using ADD is better than using curl or wget
to download a remote file, and it allows us to specify a checksum just in case.
Obviously the checksum would need to be updated when bumping Anubis version.

It's a bit awkward that we can't download, install and remove the .deb in one
single command, which I think would be more idiomatic for Docker. But we do the
install and removal in one step. Removing the file helps with the image size.

--HG--
extra : rebase_source : 7a38c92605ca3057c9a24902d00e00343109976b
extra : topic : docker-hg-apache