apache-server: run Anubis in the container
What we're doing here is, again, a bit awkward, since Docker generally suggests
to use different containers for different executables and run only one task per
container.

However, Docker has a way to run multiple tasks at the same time, see
https://docs.docker.com/engine/containers/multi-service_container/

Now users need to add `--init` argument to `docker run`, but it shouldn't be a
problem, since we list all the required commands in the README file.

We do a little dance with the hgweb.env file because Anubis assumes it's run by
systemd that has EnvironmentFile directive that reads variables from a file and
puts them into the process environment. It's *almost* the same approach that
apache2 takes with the envvars file, except in that case the variables are
actually "export"ed, so we can simply source the envvars file from our shell
and we're good to go.

Can't do the same with hgweb.env, because it doesn't export the variables, and
we can't use systemd inside the container easily, so we need to think of
something. For example, we can use a bash feature to export all variables from
the sourced files, and that's what we do. This way we achieve our goal while
maintaining the format of the hgweb.env file compatible with the "usual" way of
running Anubis, so we don't surprise the users who try to reproduce our setup
on their server.

--HG--
extra : rebase_source : 7bb5b547b2ae9bc0a988824cb96d4efc4065d5a9
extra : topic : docker-hg-apache