rust-errors: add support for backtrace for `HgPathError`
This was split off since it's a large change and it does something different.
I changed `HgPathError` to be a `struct`, and nested the original `enum` to
a `kind` field, so that we can add a `backtrace` field only once, and not
once per variant.

I then "abused" the `Into` trait to capture the backtrace at the moment the
`kind` is converted to a `HgPathError`, as explained in the implementation
of that trait. This makes it so call sites still only really need to
think about `HgPathErrorKind::*.into()` and not go through wrapping, and
saves us the trouble of setting up accessor methods that are boilerplaty
and could easily get out of sync.

--HG--
extra : rebase_source : d35aff06a70af0d8dddcb541f48846e2253e32fa
extra : topic : rust.backtrace