pyupgrade: update to Python 3.9 style
It looks like this is mostly converting from `typing` collections to builtins
and `collections.abc` collections.

This was run with a similar command as 24ee91ba9aa8 and 9d2b2df2c2ba, using
`pyupgrade` 3.20 and then running `black`:

    $ hg files -0 \
        'set:**.py - mercurial/thirdparty/** - "contrib/python-zstandard/**"' \
      | xargs -0 pyupgrade --py39-plus --keep-percent-format --keep-runtime-typing

A couple of things to note here.  First, this wasn't entirely automated, since
`pyupgrade` didn't clean up the unused imports.  `pyflakes` helped with finding
those.  Second, most of the imports that got rewritten from `typing` to
`collections.abc` (these seemed to be non-primitive types like `Iterator`,
`Mapping`, `Sequence`, etc) caused pytype (both the currently used 2023.11.21
and the latest release 2024.10.11) to emit odd pyi file content.  Specifically,
it would add a line for each `collections.abc` import like `Iterable: type`, and
then since that was nonsense, would change each usage everywhere in the file
from `Iterable` to `typing.Iterable`.  That was all reverted in order to
minimize the diffs for the pyi files being tracked in the `pytype` topic.  The
change from `typing.ByteString` to `collections.abc.ByteString` was left alone
because pytype did manage to define that correctly.  (Also, this should probably
changed to `Buffer`, if modern pytype fixes the silly issue it had with that
class[1]).

The goal in this change was to minimize the changes to the tracked pyi files
when updating to pytype 2024.10.11, since it wants to switch to using these
primitives.  That turned out to be a bust because while this does remove the
diffs for explicitly typed primitives, there are a ton of places that lack
explicit typing, and will show in a standard diff.  Still, there are a bunch of
places in the code that already used primitives for typing, so this makes the
code consistent.

In order to ignore the case changes and a few other gratuitous changes in the
diffs when jumping to pytype 2024.10.11, archive decoding can be leveraged:

    $ hg extdiff --config 'decode.**.pyi=pipe: tr "[:upper:]" "[:lower:]" | grep -Ev "(from typing import |from typing_extensions import buffer|import _collections_abc|^def (bin|hex|_pack|_unpack)[(])"'

[1] https://github.com/google/pytype/issues/1772

--HG--
extra : rebase_source : 9068db54e9b52c65b41fcc10e917e920118058ae
extra : topic : pyupgrade-39