pyupgrade: re-convert without the `--keep-runtime-typing` option
    $ hg files -0 \
        'set:**.py - mercurial/thirdparty/** - "contrib/python-zstandard/**"' \
      | xargs -0 pyupgrade --py39-plus --keep-percent-format

This is mostly about getting rid of `Optional[Foo]` in favor of `Foo | None`,
and `Union[Foo, Bar]` in favor of `Foo | Bar`.  Both of these are more readable
to me, are more concise, and don't require an import.  We already use this in
some places, so this makes the code consistent.

For some reason, this transformation changed the `Generator` in `mercurial.store`
from 3 args to 1, and that upset pytype:

    File "/home/mharbison/projects/mercurial/mercurial-devel/mercurial/store.py", line 836, in basicstore:
        Invalid type annotation 'generator[BaseStoreEntry]'  [invalid-annotation]
      generator[_T, _T2, _V] expected 3 parameters, got 1
    File "/home/mharbison/projects/mercurial/mercurial-devel/mercurial/store.py", line 863, in basicstore:
        Invalid type annotation 'generator[BaseStoreEntry]'  [invalid-annotation]
      generator[_T, _T2, _V] expected 3 parameters, got 1
    File "/home/mharbison/projects/mercurial/mercurial-devel/mercurial/store.py", line 913, in basicstore:
        Invalid type annotation 'generator[BaseStoreEntry]'  [invalid-annotation]
      generator[_T, _T2, _V] expected 3 parameters, got 1
    File "/home/mharbison/projects/mercurial/mercurial-devel/mercurial/store.py", line 977, in encodedstore:
        Invalid type annotation 'generator[BaseStoreEntry]'  [invalid-annotation]
      generator[_T, _T2, _V] expected 3 parameters, got 1
    File "/home/mharbison/projects/mercurial/mercurial-devel/mercurial/store.py", line 1187, in fncachestore:
        Invalid type annotation 'generator[BaseStoreEntry]'  [invalid-annotation]
      generator[_T, _T2, _V] expected 3 parameters, got 1

Those changes were reverted.  It can probably be converted to an `Iterable` or
`Iterator`, but that's beyond the scope of this series.

Like the previous commit, the import changes for non-primitives from `typing` to
`collections.abc` needed to be reverted here.  That was easier this time- the
command for the previous commit was run to re-add the changes and committed, the
command above was run and committed, the re-add commit was backed out, and then
the whole mess folded.  The unused imports were identified by `pyflakes` and
removed.

Note that we can't totally get away from `Optional` and `Union`.  It seems like
alias definitions and the type args when defining `attr` classes don't support
the shorthand syntax.  Because of that, the following files that were modified
by the transformation and had `Optional` and/or `Union` imports manually removed,
needed them restored:

    Optional required:
      - mercurial/cmdutil.py
      - mercurial/help.py
      - mercurial/interfaces/dirstate.py
      - mercurial/interfaces/misc.py
      - mercurial/revlogutils/deltas.py
      - mercurial/revlog.py
      - mercurial/streamclone.py
      - mercurial/utils/stringutil.py
      - mercurial/vfs.py

    Union required:
      - mercurial/bundle2.py
      - mercurial/manifest.py
      - mercurial/ui.py

After this series, the current pyi files are unchanged, except a few files that
no longer import `typing` at all.  See the previous commit for the command to
hide most trivial differences when updating to pytype 2024.10.11.

--HG--
extra : rebase_source : 05f1c160af318547abd031120d83d1a18039ec2d
extra : topic : pyupgrade-39