typing: add trivial typehints for `repository.sjoin()`
This is `store.basicstore.join()`, which doesn't have the optional extra args,
so we can trivially annotate all of this.  If it had `*args`, this would need to
be annotated with a `Callable` protocol class, given that it is an attr instead
of a method declaration.

--HG--
extra : rebase_source : 680951c7cb532fd8c19830c3a4cfb3611da792da
extra : topic : repo-join-hints