interfaces: add trivial typehints for transactions to `ilocalrepositorymain`
The `report` arg is skipped, because `ui.warn()` is assigned to it if not
supplied, which has an `*args, **kwargs` style signature.  That requires using a
`Callable` protocol class, and it's not clear yet what to call it or where it
should live.

For whatever reason, pytype 2024.10.11 stopped inferring `_transref`, and
started hardcoding it to `None`.  Put that back explicitly in order to avoid
future issues.

--HG--
extra : rebase_source : 005ac62a37966c8713f57ea520d956aff2699915
extra : topic : repo-join-hints