interfaces: add most typehints around `repository.status()`
I'm leaving out the `node1` and `node2` args for now, because I think it needs
some thought.  I see a `NodeIdT = bytes`, but there's no documentation and I'm
not sure if it's "anything that can represent a node", or the `bin(..)` form of
a node.  I think it would be good to have a way to distinguish between binary
and hex forms of a node, because there are functions that want a specific form
(though I'm not sure in practice how feasible it is to `typing.NewType()`
something for the binary form).  Let's get the easy stuff out of the way.

--HG--
extra : rebase_source : 2581844f5850f8b65e98876bd1e435745f8d4ed2
extra : topic : repo-join-hints