typing: convert from the older `istatus` to the newer `StatusT` naming style
The underlying types are identical, but do this before adding more related
types.

I left the superclass declaration alone on `scmutil.status` because unlike most
of the type aliases, this needs to be available at runtime there, instead of
just in the typing phase.  Therefore it's probably better to directly import the
module, so that 1) it stands out, and 2) we can keep the `interface.types`
module imports limited to the typing phase, in case we need to do anything
clever around types that are only available on modern pythons.

--HG--
extra : rebase_source : 8eeec4193d0e343ab4317198a4f502f3f875f6d3
extra : topic : status-hints