context: annotate `basefilectx` as having `repo` and `path` attributes
This eliminates ~20 warnings in PyCharm about fields that are accessed in the
`basefilectx`, though they weren't present (they are assigned by the subclass).
There are additional issues with `_changectx` and a handful of others, but those
are oddly handled in that they are only conditionally present, and are checked
for in `self.__dict__` before accessing (in some places, but not all).

The proper change here (and how it started out) would be to add a constructor to
the base class that accepts these args.  However, that causes an assertion when
running remotefilelog tests like this:

    [snip]
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/hgext/remotefilelog/__init__.py", line 495, in applyupdates
          return orig(repo, mresult, wctx, mctx, overwrite, wantfiledata, **opts)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/mercurial/merge.py", line 1672, in applyupdates
          for final, res in prog:
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/mercurial/merge.py", line 1470, in batchget
          wfctx = wctx[f]
                  ~~~~^^^
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/mercurial/context.py", line 91, in __getitem__
          return self.filectx(key)
                 ^^^^^^^^^^^^^^^^^
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/mercurial/extensions.py", line 528, in closure
          return func(*(args + a), **kw)
                 ^^^^^^^^^^^^^^^^^^^^^^^
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/hgext/remotefilelog/__init__.py", line 604, in workingfilectx
          return remotefilectx.remoteworkingfilectx(
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/hgext/remotefilelog/remotefilectx.py", line 501, in __init__
          super().__init__(repo, path, filelog, workingctx)
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/mercurial/context.py", line 2195, in __init__
          super().__init__(repo, path, filelog, workingctx)
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/mercurial/context.py", line 2132, in __init__
          super().__init__(repo, path)
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/hgext/remotefilelog/remotefilectx.py", line 48, in __init__
          super().__init__(repo, path, changeid, fileid, filelog, changectx)
        File "/tmp/hgtests.8njuez8_/install/lib/python3.11/site-packages/mercurial/context.py", line 1303, in __init__
          changeid is not None or fileid is not None or changectx is not None
      AssertionError: b'bad args: changeid=None, fileid=None, changectx=None'

I can't figure it out (it likely has something to do with the remotefilelog
extension multiply inheriting this class).  So rather than let the changes
following this bitrot for the 3rd time, just add the annotations and a TODO.
The generated pyi files with this and the original code with the constructor are
identical (modulo the now missing constructor line).

--HG--
extra : fix_source : 146f84164c59674cb55af181c117f5bbd333c4ab
extra : rebase_source : b665d382021bab575b3affc4714d5be342ebe778
extra : topic : context-refactor